requirements.txt
openai
streamlit
pytesseract
googletrans==4.0.0-rc1
Pillow
